refactor(api): merge integration tests for error-manager
